% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{bandwidth_nrd}
\alias{bandwidth_nrd}
\title{Bandwidth estimator for kernel density}
\usage{
bandwidth_nrd(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
estimated bandwidth
}
\description{
Bandwidth estimator for kernel density
}
\keyword{internal}
