% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{kde2d_custom}
\alias{kde2d_custom}
\title{Custom implementation of kde2d}
\usage{
kde2d_custom(x, y, h, n = 25, lims = c(range(x), range(y)))
}
\arguments{
\item{x}{x-coordinates}

\item{y}{y-coordinates}

\item{h}{bandwidth (smoothing parameter)}

\item{n}{number of grid points in each dimension}

\item{lims}{limits of the grid}
}
\value{
A list with components x, y, and z for plotting with contour
}
\description{
A simplified version of MASS::kde2d for density estimation.
}
\keyword{internal}
