% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{print.fastKPCA}
\alias{print.fastKPCA}
\title{Print method for fastKPCA results}
\usage{
\method{print}{fastKPCA}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fastKPCA}.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print method for fastKPCA results
}
\keyword{internal}
