% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppKPCA-package.R
\docType{package}
\name{RcppKPCA-package}
\alias{RcppKPCA}
\alias{RcppKPCA-package}
\title{RcppKPCA: A package for fast kernel PCA using Rcpp and Armadillo}
\description{
A computationally efficient implementation of Kernel Principal Component Analysis (KPCA) using the Gaussian (RBF) kernel. The package leverages 'Rcpp' and the 'Armadillo' C++ library for fast matrix operations to significantly speed up the kernel matrix computation, centering, and eigendecomposition steps of the KPCA algorithm. Includes methods for projecting new data points onto existing kernel principal components.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Javdat/RcppKPCA}
  \item Report bugs at \url{https://github.com/Javdat/RcppKPCA/issues}
}

}
\author{
\strong{Maintainer}: Javdat Umarov \email{lovable_residue6j@icloud.com}

}
\keyword{internal}
